backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
PyYAML>=6.0
python-multipart>=0.0.6
